requirements.txt
aiohttp>=3.9,<4.0
aiogram>=3.13,<4.0
SQLAlchemy>=2.0,<3.0
aiosqlite>=0.20,<1.0
asyncpg>=0.29,<1.0
python-dotenv>=1.0,<2.0
